model refactoring-> classes and packages
