+++ src/main/java/org/fundacionjala/converter/controller/MetadataController.java
@@ -12,8 +12,8 @@ import org.fundacionjala.converter.controller.request.RequestMetadataParam;
 import org.fundacionjala.converter.controller.request.RequestMetadataValidator;
 import org.fundacionjala.converter.model.ChecksumMD5;
 import org.fundacionjala.converter.model.entity.File;
-import org.fundacionjala.converter.model.metadata.MetadataExtractor;
-import org.fundacionjala.converter.model.metadata.MetadataParam;
+import org.fundacionjala.converter.model.command.metadata.MetadataExtractor;
+import org.fundacionjala.converter.model.command.metadata.MetadataParam;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.RequestMapping;
