--- src/main/java/org/fundacionjala/converter/model/Param.java
@@ -1,4 +1,0 @@
-package org.fundacionjala.converter.model;
-
-public abstract class Param {
-}
