+++ src/main/java/org/fundacionjala/converter/model/command/CommandBuilder.java
@@ -6,7 +6,7 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package org.fundacionjala.converter.model;
+package org.fundacionjala.converter.model.command;
 
 import java.util.List;
