+++ src/main/java/org/fundacionjala/converter/model/command/ConvertDoc.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.converter.model;
+package org.fundacionjala.converter.model.command;
 /**
  * Copyright (c) 2020 Fundacion Jala.
  *
