+++ src/main/java/org/fundacionjala/converter/model/command/ICommand.java
@@ -1,0 +1,8 @@
+package org.fundacionjala.converter.model.command;
+
+import java.util.List;
+
+public interface ICommand {
+
+    List<String> createCommand();
+}
