+++ src/main/java/org/fundacionjala/converter/model/command/Param.java
@@ -1,0 +1,4 @@
+package org.fundacionjala.converter.model.command;
+
+public abstract class Param {
+}
