+++ src/main/java/org/fundacionjala/converter/model/command/metadata/MetadataParam.java
@@ -6,8 +6,8 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package org.fundacionjala.converter.model.metadata;
-import org.fundacionjala.converter.model.Param;
+package org.fundacionjala.converter.model.command.metadata;
+import org.fundacionjala.converter.model.command.Param;
 
 /**
  * @author Angela Martinez
