+++ src/main/java/org/fundacionjala/converter/model/commons/ChecksumMD5.java
@@ -6,7 +6,7 @@
  * Information and shall use it only in accordance with the terms of the
  * license agreement you entered into with Fundacion Jala
  */
-package org.fundacionjala.converter.commons;
+package org.fundacionjala.converter.model.commons;
 
 import java.io.File;
 import java.io.FileInputStream;
