--- src/main/java/org/fundacionjala/converter/model/parameter/ffmpeg/FfmpegParameter.java
@@ -1,7 +1,0 @@
-package org.fundacionjala.converter.model.parameter.ffmpeg;
-
-import org.fundacionjala.converter.model.parameter.ModelParameter;
-
-public abstract class FfmpegParameter extends ModelParameter {
-
-}
