+++ src/main/java/org/fundacionjala/converter/model/parameter/multimedia/AudioParameter.java
@@ -1,8 +1,8 @@
-package org.fundacionjala.converter.model.parameter.ffmpeg;
+package org.fundacionjala.converter.model.parameter.multimedia;
 
 import java.util.List;
 
-public class VideoParameter extends FfmpegParameter {
+public class AudioParameter extends MultimediaParameter {
 
     /**
      * List<String> getParameter()
