+++ src/main/java/org/fundacionjala/converter/model/parameter/multimedia/MultimediaParameter.java
@@ -1,0 +1,7 @@
+package org.fundacionjala.converter.model.parameter.multimedia;
+
+import org.fundacionjala.converter.model.parameter.ModelParameter;
+
+public abstract class MultimediaParameter extends ModelParameter {
+
+}
