+++ src/main/java/org/fundacionjala/converter/model/parameter/multimedia/VideoParameter.java
@@ -1,8 +1,8 @@
-package org.fundacionjala.converter.model.parameter.ffmpeg;
+package org.fundacionjala.converter.model.parameter.multimedia;
 
 import java.util.List;
 
-public class AudioParameter extends FfmpegParameter {
+public class VideoParameter extends MultimediaParameter {
 
     /**
      * List<String> getParameter()
